+++ gen/plugin.go
@@ -216,8 +216,12 @@ func (g *generateServiceBuilder) buildFieldGroup(fs compile.FieldGroup) ([]*api.
 			return nil, err
 		}
 
+		name, err := goName(f)
+		if err != nil {
+			return nil, err
+		}
 		args = append(args, &api.Argument{
-			Name: goCase(f.Name),
+			Name: name,
 			Type: t,
 		})
 	}
@@ -250,9 +254,13 @@ func (g *generateServiceBuilder) buildType(spec compile.TypeSpec, required bool)
 		if err != nil {
 			return nil, err
 		}
+		name, err := goName(s)
+		if err != nil {
+			return nil, err
+		}
 		t = &api.Type{
 			ReferenceType: &api.TypeReference{
-				Name:       goCase(s.Name),
+				Name:       name,
 				ImportPath: importPath,
 			},
 		}
@@ -320,10 +328,15 @@ func (g *generateServiceBuilder) buildType(spec compile.TypeSpec, required bool)
 			return nil, err
 		}
 
+		name, err := goName(s)
+		if err != nil {
+			return nil, err
+		}
+
 		return &api.Type{
 			PointerType: &api.Type{
 				ReferenceType: &api.TypeReference{
-					Name:       goCase(s.Name),
+					Name:       name,
 					ImportPath: importPath,
 				},
 			},
@@ -335,9 +348,14 @@ func (g *generateServiceBuilder) buildType(spec compile.TypeSpec, required bool)
 			return nil, err
 		}
 
+		name, err := goName(s)
+		if err != nil {
+			return nil, err
+		}
+
 		t = &api.Type{
 			ReferenceType: &api.TypeReference{
-				Name:       goCase(s.Name),
+				Name:       name,
 				ImportPath: importPath,
 			},
 		}